requirements.txt
requests
colorama
httpx
flask
celery
pyyaml